[MOD-2100] Added 64-bit support
Specifically:

* Updated Xcode project settings to support 64-bit architectures
* Updated MagTek library to V114.01.
* Updated build script to check for architecture, etc.
* Changed module version from 2.0.0 to 2.1.0

The MagTek library comes in two pieces, one for devices and one for the simulator. For
sake of simplicity, these two are merged into one file when used by the module. This
technique has been used in the previous version of the module; this commit just updates
the combined file. See MagTek Documentation/Lib/README.txt for more detail.

The library update includes a slightly modified header file, indicating some changes in the
library interface. The module doesn't seem to be referencing any of the changed parts, so
the module code itself is not being modified by this commit.

Note that the somewhat misleadingly named "MagTek Documentation" folder contains everything
in the original MagTek SDK package, including the library itself and a sample app (as well as
documentation); however, it is not actually used in the module for anything. The header and
library files used by the module are in the 'lib' folder. This commit merely updates the existing
"MagTek Documentation" folder's contents, it does not create it.